cloud-function/requirements.txt
google-cloud-bigquery==2.23.2